backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.13.0
django-filter==23.4
psycopg2-binary==2.9.3
python-dotenv==1.0.0
djoser==2.1.0
gunicorn==20.1.0
webcolors==1.11.1
flake8==6.0.0
flake8-isort==6.0.0
reportlab==4.0.7
django-colorfield==0.11.0